Separate unit tests into different files
